refactor: move sample promoter arrays to fixtures
