+++ lib/fixtures/promoter-profile.ts
@@ -1,0 +1,17 @@
+export const sampleEmployers = [
+  { value: "agency_alpha", label: "Agency Alpha Solutions" },
+  { value: "beta_workforce", label: "Beta Workforce Inc." },
+  { value: "gamma_talents", label: "Gamma Talents Co." },
+]
+
+export const sampleClients = [
+  { value: "client_mega_corp", label: "MegaCorp International" },
+  { value: "client_tech_innovators", label: "Tech Innovators Ltd." },
+  { value: "client_global_retail", label: "Global Retail Group" },
+]
+
+export const promoterStatuses = [
+  { value: "active", label: "Active" },
+  { value: "inactive", label: "Inactive" },
+  { value: "suspended", label: "Suspended" },
+]
