+++ lib/promoter-profile-schema.ts
@@ -39,22 +39,3 @@ export const promoterProfileSchema = z.object({
 })
 
 export type PromoterProfileFormData = z.infer<typeof promoterProfileSchema>
-
-// Sample data for dropdowns
-export const sampleEmployers = [
-  { value: "agency_alpha", label: "Agency Alpha Solutions" },
-  { value: "beta_workforce", label: "Beta Workforce Inc." },
-  { value: "gamma_talents", label: "Gamma Talents Co." },
-]
-
-export const sampleClients = [
-  { value: "client_mega_corp", label: "MegaCorp International" },
-  { value: "client_tech_innovators", label: "Tech Innovators Ltd." },
-  { value: "client_global_retail", label: "Global Retail Group" },
-]
-
-export const promoterStatuses = [
-  { value: "active", label: "Active" },
-  { value: "inactive", label: "Inactive" },
-  { value: "suspended", label: "Suspended" },
-]
